--- a/src/database/index.ts
+++ b/src/database/index.ts
@@ -1,0 +1,2 @@
+export * from './SpurDatabase';
+export * from './SyncManager';