… agent-assignments/crush-agent-tasks.md
… crush-algorithm-optimization-results.json
… docs/agent-activation.md
… docs/agent-communication.md
… docs/agent-system.md
… docs/execution-plan.md
… docs/optimized-algorithms.md
… docs/phase2-steel-integration.md
… external-agent-coordination.json
… phase-planning-coordination.json
… src/accessibility/config.ts
… src/accessibility/tester.ts
… src/assistant/core/context-aware-response.ts
… src/assistant/core/index.ts
… src/assistant/inboxzero/index.ts
… src/assistant/leon/index.ts
… src/assistant/mailvelope/index.ts
… src/assistant/ml/auto-optimization.ts
… src/assistant/ml/monitoring.ts
… src/assistant/ml/optimizer.ts
… src/assistant/nlp/advanced-nlp.ts
… src/assistant/nlp/index.ts
… src/assistant/proactive/index.ts
… src/assistant/skills/index.ts
… src/assistant/voice/index.ts
… src/assistant/voice/privacy-voice.ts
… src/assistant/vscode/index.ts
… src/capture/collectors/base.ts
… src/capture/collectors/browser.ts
… src/capture/collectors/manager.ts